feat: create NavMenu component and integrate it into Header
